Replace the 'required' parameter by using 'default'=(anything) wherever one would have previously used 'required'=false.  This removes a previous need for annoying boilerplate config to explicitly set 'required'=false when setting a default.  Use 'default'=null when a parameter is completely nonessential and your program does not need to expect any value.
This also means that if parsing completed, a key in the results array will be defined for every parameter in your configuration -- you no longer need to prefix all your $results accesses with '@' in the case of parameters that are not required and do not have a default value.

Also, 'type'=bool is trivially 'default'=false (and 'required'=false, if you prefer to think of it that way).