Remove the configuration to accumulate instead of throw errors from the test cases, since PHP/PHPUnit can't support what seems like it should be a fairly simple comparison without segfaulting.
Incidentally, this is also now one of those cases that makes me unhappy with all the xUnit family spawned from the SUnit of yore, and what made me make a JUnit alternative in java.  What I'd really like to do is make an assert that the count of the errors array is 0, and if that fails, keep going long enough to loop over it and print a terse summary of the exception messages, or rethrow one of them... but that's impossible, since the first assert to fail will deprive you of the opportunity to do any further examination of your situation.